Implemented the add to cart feature

+++ ACETutoring/store/views.py
@@ -32,44 +32,40 @@ def books(request):
 
 def bookdetail(request, pk):
     item = get_object_or_404(RevisionBook, pk=pk)
-    related_items = RevisionBook.objects.filter(category=item.subject).all()
-    return render(request, 'books/bookdetail.html', {
+    related_items = RevisionBook.objects.filter(subject=item.subject).exclude(pk=pk)[0:3]
+    return render(request, 'store/bookdetail.html', {
         'revisionbook': item,
         'related_revisionbooks': related_items
     })
 
+def add_to_cart(request, revisionbook_id):
+    product = RevisionBook.objects.get(id=revisionbook_id)
+    cart_item, created = CartItem.objects.get_or_create(product=product, user=request.user)
+    cart_item.quantity += 1
+    cart_item.save()
+    return redirect('store:view_cart')
 
-"""
-# enumerate all the books
+
 def book_list(request):
     revisionbook = RevisionBook.objects.all()
     return render(request, 'frontpage/bookindex.html', {'revisionbook': revisionbook})
  
-# view the books currently in the cart  
 def view_cart(request):
     cart_items = CartItem.objects.filter(user=request.user)
-    total_price = sum(item.product.price * item.quantity for item in cart_items)
-    return render(request, 'frontpage/cart.html', {'cart_items': cart_items, 'total_price': total_price})
- 
-def add_to_cart(request, book_id):
-    product = RevisionBook.objects.get(id=book_id)
-    cart_item, created = CartItem.objects.get_or_create(product=product, 
-                                                       user=request.user)
-    cart_item.quantity += 1
-    cart_item.save()
-    return redirect('cart:view_cart')
- 
+    total_price = round(sum(item.product.price * item.quantity for item in cart_items),2)
+    return render(request, 'store/cart.html', {'cart_items': cart_items, 'total_price': total_price, 'checkout': False})
+  
 def remove_from_cart(request, item_id):
     cart_item = CartItem.objects.get(id=item_id)
-    cart_item.delete()
-    return redirect('cart:view_cart')
-
-"""
-
-
-
-
-
-
-
+    if cart_item.quantity > 1:
+        cart_item.quantity -= 1
+        cart_item.save()
+    else:
+        cart_item.delete()
+# due to the scope of the project, I will not be handling payment so this just resets the queue 
+def check_out(request):
+    cart_items = CartItem.objects.filter(user=request.user)
+    for item in cart_items:
+        item.delete()
+    return render(request, 'store/cart.html', {'checkout':True})
